refactor: 디렉토리 경로 수정 #143
