+++ src/app/chat/[id]/_shared/components/chat-input/chat-input.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState } from 'react';
 
-import CancelProgressModal from '@/app/chat/[id]/_shared/components/cancel-progress-modal/cancel-progress-modal';
+import CancelProgressModal from '@/app/chat/[id]/_shared/components/chat-input/cancel-progress-modal/cancel-progress-modal';
 import {
   usePatchCancelProgress,
   useSendChatMessageImage,
