+++ src/app/chat/[id]/_shared/components/chat-message-list/fulfillment-message-card/fulfillment-message-card.module.scss
